Playing around with adding "default" to the cobalt_client package.
I am unable to get cobalt_testapp to pass on the CQ for
https://fuchsia-review.googlesource.com/#/c/cobalt_client/+/58375/.
The test that attempts to use the instance of Cobalt from the
environment is failing and I don't understand why. This test
passes fine on my computer so I am experimenting with adding
dependencies to see if that helps.

Change-Id: Ic829e4bbb4d69cc8758d260483253499edb204cb