Use an even more robust termination criterion.
It still is not enough to just wait for conflict resolution to
notify watchers. They can still be notified of the value pre
resolution. That's a property of the Ledger API right now.

Therefore, we just wait for the notifications of the resolved
values. This has two disadvantages: (1) it makes the termination
condition redundant with the tested expecations, and (2) failure
takes longer because it has to timeout. However, it's currently
the only way to determine how long to wait because the actual
condition to wait for (that a specific conflict was resolved and
the watchers notified of it) is not observable in the Ledger API.

Enable disabled test cases.

FW-278 #done again
LE-309 #comment

Change-Id: I0547e7a816385027033cd8dff1d7bde1799f74f6