[arm64] Check boot EL2 when creating a guest.
This mirrors the VMX check in the x86, and helps us to early exit when
creating a hypervisor guest if it's not supported on the device.

Also split the hypervisor.cpp file into guest.cpp and vcpu.cpp. It's
still a skeleton, but it helps while we flesh it out.

Change-Id: I482f55b5a448c6b26acb363ec3bad76a9ed8c5dc