[kernel][x86] Add a global variable to cache x86_feature_test(X86_FEATURE_SMAP)
This simplifies the check in _x86_copy_to_or_from_user().  The
assembly code can access the global rather than having to take
smap_avail as an argument.

This is more of a cleanup than a performance improvement.  Though
x86_feature_test() isn't trivial, it doesn't appear to add noticeable
overhead to the context switch.

Change-Id: I93529a0093366a212ae9214ac0e0ffc886ff1595