Extract CloudDeviceSet into a separate library.
This patch extracts the implementation of cloud device set into a
separate library, so that it can be used both in the ledger binary and
in the WIP standalone cloud provider.

After the switch we will move it under /cloud_provider_firebase.

LE-307 #comment extract CloudDeviceSet into a separate library

Change-Id: I58debc805a57c3176ed4b635eca565daa0be08c4