Remove the thread-per-output in the mixer.
Remove the dedicated thread-per-output which had been previously used.
Instead, switch over to using timers from the recently refactored
audio dispatcher API.  Also, depend on the dispatcher framework's
ExecutionDomain serialization behavior to protect the majority of the
internals of the output's mixer state instead of needing to hold onto
an output-wide mutex for most of the mix operation.

Change-Id: I71e105c9327e63849dbd466b49e4f6134d351d84