[entropy] Enable jitterentropy as entropy source
Testing shows that, with the newly set parameters (bs=64, bc=512, ml=32,
ll=1, raw=true), each byte of data contributes approximately 0.58 bits
of min-entropy on the rpi3 and 0.5 bits on qemu-arm64.  A safety factor
of 0.9 gives us 0.50 * 0.9 * 1000 == 450 bits of entropy per 1000 bytes
of random data.

This adds about 40ms to boot on rpi3 and 30ms on qemu-arm64.
Jitterentropy doesn't work on x86 as of this commit (the timer isn't
ready during early boot), so there should be negligible slowdown on x86.

On rpi3, about 10ms is due to the actual entropy draw from
jitterentropy, and 30ms is due to the startup testing built into
jitterentropy. Once we have entropy source testing in the kernel, we can
replace the jitterentropy testing time to reduce blocking during the
single-threaded, pre-VMM boot phase.

Change-Id: Ibf258412c7947e4e828926a88c0897f22199d7d3