Add new install_host_tools template in //build/host.gni
This declares that some particular host_toolchain targets should be
copied into a "host tools" directory that is appropriate e.g. for a
developer to add to their shell's $PATH.  This segregates host tools
that exist purely to be used by GN rules in the build from things
that are tools used outside the build by developers (or scripts that
are not part of the GN build itself) for testing, debugging, etc.

Change-Id: Ic0f4bcdc02f8905792d5bfc3cb80e83aa18543f2