[mtx_t][threadsafety] Add thread safety annotations to mtx_t
This adds locking annotations to mtx_lock and mtx_unlock. This means
that functions which return holding more locks than they started with
(or with fewer) must get annotations. A few cases in which locks are
conditionally acquired is beyond the analysis, and marked as such.

This does _not_ mean that any data is marked as guarded by a mtx_t,
but it does enable that.

Change-Id: Ia545f1515e238ab690888f98b84b7392ba809c4f