[minfs] Fix inode consistency state during write error
If a write fails (for example, due to low disk space),
the inode size should only be updated to include
all successful writes, not those that did not get
a chance to propagate to disk.

Simplified the MinFS WriteInternal lifetime a little
(by forcing the function to complete through 'done')
and add a test for this condition.

Change-Id: I898111b47c4afd90b5a4197acbf6f12bf9b38009