[kernel][handles] Use fbl::atomic<> for the process_id_ field
This field can be accessed concurrently, so we should make it atomic
to avoid undefined behaviour in C++.  It is possible for one process
to read this field (e.g. in GetHandledLocked()) while other process
writes to it (e.g. in RemoveHandleLocked()), because the field is only
guarded by the per-process handle_table_lock_.  The two processes may
be holding different instances of handle_table_lock_.

In practice this won't affect the generated code, because we use
memory_order_relaxed.

Change-Id: Idf4a3e7635c14cf28aa68acb3893646855cda85b