[guest] Add clocksource=tsc to fix linux timer.
Linux can still mark it as unstable but is forced to use it anyway.

Setting CONFIG_HZ_100=y on the Linux compile reduces the drift and
makes Linux happier with tsc, so presumably we aren't sending our
timer interrupts with the frequency that the watchdog would like.

Experiments with kvm and disabling the fancier timers (acpi_pm, hpet)
that we don't implement doesn't introduce drift, so they're presumably
handling this in some nicer fashion.

For now, I'll change the Linux kernel config too and we can revisit this
later.

Change-Id: I7bcdc694ae61f9995097ad06385a2aa521e5d6ae