Fixes blocked ports in open wlan's.
3 problems:
1) We checked for null rather than empty() when making a port decision.
Since we always initialize the Rsn array, a null check always failed.
Originally, the idea why the array is always initialized is to not
constantly allocate memory on each beacon frame.
2) We initialize the array with Rsn::kMaxLen to set its capacity to the
maximum required size to not have smaller allocations when writing the
actual RSNE content. However, this caused the array to have a size() of
Rsn::kMaxLen and not just a capacity of that value. Hence we have to
resize(0) after initializing. Since at least in theory, the RSNE could
change from beacon to beacon frame, we also have to resize(0) the rsn
when resuing the struct.
3) We are using FIDL rather than something C++ specific. porce@ will
address this part of the issue.

Change-Id: Ia2e2c545aff589d1af799f1a76cf9e2fc593f781