Don't forward deps into the non-shared shared library target
If the shared_library has a dep on a source_set then the source_set
files were being included into executables that depended on the
shared_library. Removing deps should fix that, and shouldn't break other
things because public_configs are passed through and a public_dep is
added to the real shared library.

Change-Id: Ied9cc6f6a030081675ad41fbef1c053e525b7bf4