[dev][platform-bus] Add support for dynamically enabling and disabling platform devices
The platform bus implementation driver can now enable or disable a platform device
using the new pdev_device_enable() API. This will add or remove the DDK device for a
platform device. A new "enabled" flag in the MDI can be used to specify that a
platform device not be enabled at startup.

We will use this to dynamically add and remove the platform device for XHCI when
switching between USB host and peripheral roles.

Change-Id: I126ea70b0dba6573230528899a00fd21320ee6ea