[download] Remove directory we will unzip into, not subdir
This changes the download behavior for zips to always nuke the
directory the archive will unzip into before unzipping instead of
trying to delete a subdirectory of the target file. Since our zip
archives are downloaded into distinct subdirs, this ensures the
downloaded directory matches the archive contents (only) after an
update.

This mitigates some issues with the Go build that manipulate the
directory in //buildtools after downloading to make sure that the local
changes are nuked after updates.

Change-Id: I008c18af7013cce095511cbc4a6c889fa9561558