Timeout Teardown() calls.
The class AsyncHolder holds a pointer to an instance of a class
with an asynchronous Teardown() method. AsyncHolder has a method
Teardown() on its own, which invokes the underlying Teardown()
bounded by a timeout. This is analogous to how AppClient invokes
the Terminate() method of its application instance.

Member os user runner with a Teardown() are now terminated through
AsyncHolder, with an explicit timeout.

The component_context test, which includes an "unstoppable agent"
now hots the agent runner timeout, no longer the user runner timeout:

[INFO:user_runner_impl.cc(283)] UserRunner::Terminate()
[INFO:agent_runner.cc(52)] AgentRunner::Teardown() 4 agents
[WARNING:apps/modular/lib/fidl/app_client.cc(94)] AppTerminate() timed out for file:///system/apps/modular_tests/component_context_unstoppable_agent
[WARNING:apps/modular/lib/common/async_holder.cc(29)] Teardown() timed out for AgentRunner
[INFO:user_runner_impl.cc(354)] UserRunner::Terminate(): done

It is not clear why we hit the AgentRunner timeout even though
we already hit the timeout for the Agent app. To be followed up.

FW-295 #done

Change-Id: I4baa19fa30edc1f320af86525b4faeae3efdbd4a