[hypervisor] Add ZX_GUEST_TRAP_BELL.
This new trap type allows a trap to be generated when a memory region
is accessed. Unlike ZX_GUEST_TRAP_MEM, this does not attempt to fetch
the instruction that generated the access. We can also deliver the
trap asynchronously via a port, as we do not need to pause execution
of the VCPU that caused the trap.

In a follow up to this change, I'll begin explicitly marking all the
memory regions that require a trap, which will also allow us to
optimise our page-fault path.

Change-Id: I0505dfe79abf997b0544ff6136bb739e52f8ba96