Explicitly coodinated timeouts.
All timeouts are explicit and defined in one location, so it's
easier to keep them consistent. The eventual plan is still to
have them negotiated.

Also, all AppClient timeouts are now explicit, so it's easier to
not forget about them.

Delete the AppTerminate() wrapper, instead we call Teardown()
everywhere.

There is still an odd discrepancy between AsyncHolder.Teardown()
and AppClient->Teardown(). We resolve this in a followup.

FW-295

Change-Id: I375a728d53ef19849833f911963d5dfa3a60eef0