[fs] Move VFS RPC code into Connection object.
The RPC functions become members of the Connection object itself
such that the Connection encapsulates both the state and the behavior.

As a result of this change, we no longer need separate AsyncDispatcher
or AsyncHandler objects since all of the work is being done by the
Connection.

Also moves the readable/writable predicates to a common location
and fixed them to use O_ACCMODE, which will make it easier to add
support for O_PATH later.

Change-Id: Ib2940aae1ff03d36d3310d6887ce725069373030