Fix IncrementalChangeCall.
The call to IncrementalWriteCall was not pegged to the flow token,
so the operation would not wait until it finishes. Once we switch
all operations to be invoked asynchronously, in upcoming change
https://fuchsia-review.googlesource.com/#/c/peridot/+/66807/,
the write operation would never run because it gets deleted before
it's invoked.

Unittests still pass, as nothing has changed. They will need to
be adjusted once the invocation order changes.

Change-Id: I909af99d6ce6057fff77bd0d3878d0896d6775d2