[modular] StoryStorageImpl -> LinkImpl
Move Link related storage operations from StoryStorageImpl
to LinkImpl.

This is not just refactoring, but a semantic improvement:

* Story page operations now run concurrently for each link, and
  are no longer serialized across all links.

* On teardown of a story, we already act as if this would be the
  case, by calling Sync() on every link rather than just on
  StoryStorage. Removed the corresponding TODO to only Sync()
  on StoryStorage.

* We now use PageClient notifications for a key prefix in a novel
  way: for dynamically computed prefixes (see also below).

Other improvements:

* There is now one mechanism less to propagate ledger updates
  to the data consumers, as LinkImpl is now simply a PageClient,
  rather than using its own notification and registration
  protocol.

* The LinkImpl life cycle is now simpler because it doesn't need
  to unregister itself in its destructor.

* It is now easier to add watcher updates for other data
  types than links stored in the story page. (This is done in
  a followup CL.)

To support testing, add a hook to OperationBase that
notifies an observer function of every Operation instance
executed. Together with watching the story page through
a ledger client peer, this yields the same information as mocking
the LinkStorage API used to give. The link impl test now uses
operation counts obtained that way to assert upper bounds on
the number of expensive ledger storage operations per Link
operation.

The prefix passed to PageClient is now passed as string with
a default rather than as const char*, because the prefix is
no longer always a constexpr computed at build time. In a LinkImpl
page client, the prefix to watch is computed at runtime from story
ID and link ref. The empty prefix is now specified by an empty
string (or omitting the argument) rather than by a nullptr.

TODO in followup: StoryStorageImpl is still used to store other
story related data that are under the management of
StoryControllerImpl. We next move the remaining operations
there, turn it into a PageClient. This makes changes in those
data observable too, and removes the strange duplicate
OperationQueue in StoryControllerImpl/StoryStorageImpl.

FW-318 #done

Change-Id: I4d34f0f3d7b8a53a7ae13fdda99bb2137ff46898