Fix StoryProvider.GetLinkPeer().
After the introduction of LedgerClient, link peers were created with
a Page connection to their story page shared with other peers, rather
than with a dedicated separate Page connection. This defeats the
purpose of Link peers.

In order to set up a dedicated Page connection with a PageClient, a new
LedgerClient of the same Ledger must be created, which then creates new
Page connections for its own PageClients. The method GetLedgerClientPeer()
added to LedgerClient creates this separate LedgerClient, and is used in
the GetLinkPeer() implementation.

FW-277

Change-Id: I0ff00cfaade0e2497b98e0254162166950bf411e