[zircon][hypervisor] Fix SCR_EL3.HCE offset and set SCR_EL3.NS flag
According to the ARM Architecture Reference Manual (ARMv8) correct bit
for SCR_EL3.HCE is 8 and not 1. Also EL2 is only available in non-secure
work, so we need to set SCR_EL3.NS too.

Note that I currently don't have an equipment to test it on a real
hardware, so it's purely documentation based fix.

Change-Id: I629ad6765f670e5ae8552a91b467258f288f77f5