[sdk] Clean up path handling, make failures nonfatal
As a convenience makesdk tries to make the path entries in the ids.txt
for kernel debug info relative. This fails on the bot, likely due to
non-canonical paths. This makes the handling a bit more robust and
allows the SDK packaging to continue even if it can't figure out how to
turn some paths into relative paths.

Change-Id: I83e804e1523fe4ac60abb6bb8285d80478b47b94