[hypervisor][lib] Fix BlockingPortAllocator.
Because of the way we Alloc first before we Wait on the semaphore in
order to reuse the semaphore's internal count, we can actually overrun
the Arena and cause issues. To alleviate this, reduce the semaphore's
internal count by 1. We waste on PortPacket worth of memory, but we
don't need to keep an external count and use additional locking.

Change-Id: Ica94385f35cd74c10d24c8ee3bf08f74bc1a1620