[dash] Fix error reporting from process_launch
Before this, an int in the style of unix process exit() values (0
success, low numbers or high 120s different kind of bad, etc.) was
returned, but logged as though it were a zx_status_t. This tweaks
process_launch to return both styles and to print the more informative
zx_status_t to stderr, with $? still set correctly to the unix style.

Change-Id: I66cb615b2b83318829029f506f1e0f8156051eb4