Fix test harness.
The test harness needs to create an ApplicationContext to send results.

Until now, it was possible to create one without a message loop.
Following https://fuchsia-review.googlesource.com/c/garnet/+/66588,
creating an application context now requires one.

This change test reporting so that results are recorded, and are only
sent at the end so that we only need to create a message loop after all
the tests have been run.

Change-Id: I6dfe8424bd05b2166c9dcea4b39d6d5d52cefdc4