[modular] Change TestWithMessageLoop to create a new message loop per test
This is now necessary because the test harness needs to create a message
loop when all tests are done, and this is not possible if one is left
around by the tests. Moreover, it ensures a better isolation between
tests.

The issue that prevented this before was that the
LedgerRepositoryForTesting was using a singleton for ApplicationContext
that was bound to the first run loop created.

This CL also modifies LedgerRepositoryForTesting to use a new directory
each run to ensure a clean ledger.

Change-Id: I8c8f2144cf8ad06ffe65f03dd9872785a122c2da