Fix Zircon tests
For tick_length, zx_nanosleep makes no guarantees about an upper length of the sleep so this test shouldn’t try to test for one.

For socket_basic, the return code for this error state appears to have changed.

These tests seem questionable to me, as they appear to be testing Zircon, not the Rust wrapping of Zircon.

Change-Id: Iaf09aff61874868efefe7ea785e87ac2cf508521