netconnector: recheck NICs when address family is unspecified
netconnector's mDNS implementation has code to repoll GetInterfaces
when nics don't appear to be properly warmed up. This logic was
missing for the case when a nic's address family is unspecified.
That particular code block also emitted a scary log message when
the behavior is actually expected.

This CL also filters out loopback addresses.

Change-Id: I6a563cb2d34405554c2a47b1ae5a1d73a850c2c3