[context] Add ContextValueWriter interface & impl.
There are some odd race-conditions with ScopedContextValue, the
client-side convenience interface to tie the lifetime of a context value
to a local variable. A better approach is to tie the lifetime to a FIDL
handle explicitly. There are no race conditions this way.

Everything we store in the context service is actually ephemeral, and
the lifetime should be tied to the item it represents elsewhere in the
system (the module, story, entity, link, etc). This is an easy way so
that, at least in C++, it's as easy as destroying the proxy object and
closing the channel.

This also makes it easier to interface with the service at all: you
don't need to wait for an Add*() call before you can make updates, and
we can take advantage of interface request pipelining to streamline the
client-side code a bit.

Original CL: https://fuchsia-review.googlesource.com/#/c/maxwell/+/65611/

MW-255

Change-Id: I069c362503e1869397deda6601f02628e3fca1fe