Make LinkImpl test not depend on call ordering.
The test checked on two things: (1) calls to StoryStorage were
observed in the mock implementation, (2) calls to LinkWatcher
notifications were observed in a LinkWatcher implementation.

The problem was that there were assertions about story storage
observations made in the LinkWatcher implementation. Thus, the
test relied on the storage calls being observable by the time
of the watcher notifications, hence that the two occur in a
prescribed order.

Moreover, although we don't rely on the actual order, the order
that was prescribed here is the least favorable one: Storage
is first, but we want LinkWatcher notifications to occur as
fast as possible. So if we had a preference at all, we would want
LinkWatcher to be called first, Storage then.

Solution is to change the test to rely on story storage change
writes to trigger validation, and just insert additional validation
into the link watcher that ensures it's not called excessively.

Change-Id: I9ba2da9fd2f69cb6c1e1095ba0222ddcd73f7bf8