[hypervisor] Disable VMX in VmxCpuState destructor only if it was fully enabled
If enabling VMX failed for any CPU, VmxCpuState::Create will queue a task
to disable it on any CPU that already enabled it. The VmxCpuState object
was already created and will try to disable it for all CPUs on
destruction.

This bug was introduced by 776d4eed1b. On my machine, runtests with kvm
enabled fail with a crash while executing vmxoff.

Change-Id: Ie37f30c73017b3069189032b610ad28d16d517e3