[auth] Add support for token cache.
This cl adds support for caching short lived oauth tokens in a in-memory
LRU based cache. This is acheived by having two data structures - an
unordered map and a list.

The list keeps track of the key |CacheKey| along with the value of corresponding
|OAuthTokens|. List can hold cache entries upto a max of cache_size as specified
by user during initialization. Map is indexed by |CacheKey| and points
to the beginning of the list.

Get,Put,Delete,and HasKey methods are provided that work on a given CacheKey.
AddFirebaseToken method either adds a new firebase token or updates an existing
firebase token for the given cache entry.

Change-Id: Ia8c7130c2a76d336cf915253df5ea6d0106a5a31