media: add release handler to video renderer view to fix leak
The video renderer view needs to register for a callback, which is
basically just set_connection_error_handler for the view listener.
It deletes itself in that callback.

This fixes a memory leak in media_service. The view holds a reference
to a VideoFrameSource, which integrates a VideoConverter, which has
a multi-megabyte lookup table.

Change-Id: I7578f130d868450e9e5ac403380ca23871c42fad