[rust] Split client and server FIDL traits for honest error handling
While FIDL clients can reasonably return errors to be handled by the
user, FIDL services don't currently have an out-of-band way to return
protocol errors or other failure messages. When a server is unable to
process a request but a response is required, the only way to indicate
failure to the client is to close the channel. When no response is
required, the server can choose to either swallow the error or close
the channel.

Change-Id: I244e11d351c2ad130d582354632e75897634c9ab