[devcoord][devhost] rearrange how we bootstrap /dev/sys
The devcoordinator now requests that a specific driver
(acpi or platform-bus) is invoked to create() this device
rather than playing games with the binding process.

This simplifies the process of passing startup data to
/dev/sys, and communicating with it later (for shutdown,
etc).

The zx_device_t no longer needs to cache the args and
we can remove device_get_args()

Change-Id: I4c7e2a417acfb4e6c3841356a2b3b62cc8e8df80