[hypervisor][ulib] Use dynamic mapping of IO regions.
Add a generic IoHandler interface to allow devices to provide callbacks
to regions that they have set traps for. Set the address of the mapping
object as the trap 'key' so that we can reuse the kernel trap
resolution logic to select the appropriate userspace handler.

This implementation also consolidates the ports and worker threads used
for handling async operations into a single port with a fixed size thread
pool for all devices.

Add an implementation of the UART to utilize the new interface.

Change-Id: I88a8ced3ecb4d933d8ecde0970892da16d0e80e8