[fvm] Fix random-op test race condition
When initializing threads, the "parent" thread
launches an array of thread [0, 1, ..., ThreadCount).

When the child threads initialize, they grabbed
a "thread count" number to access their index into
the parent's array. Although the access to this thread
count variable was locked, it was not ordered between
threads. As a consequence, the "0th index thread" could
end up grabbing "tid = thread_count != thread_index".
Since the parent thread reads/writes to each thread's
private data (vpartition) after join, this may result
in the parent thread altering the state for a test
thread that is still executing, which may result in a
failure.

To fix this issue, this patch provides each thread with
an auxilliary argument, "thrd_args_t", which provides
the thread ID up-front, as well as a pointer to the
globally shared state.

Change-Id: I4843d6c93a40f7a1f17e78b48e6ad9ebbecea7ba