[fidl] Fail fast for C++ serialization problems
FIDL serialization only fails because of programmer error. The C++
bindings are vulnerable because it's easy to fail to initialize a
required field in a struct. The current behavior is that in debug builds
you get a warning when sending the message but often a failure & crash
on the receiver (especially if it's Dart).

This change makes serialization failure FATAL rather than WARNING in
debug builds. This means the developer gets a stack trace that's more
likely to be near their error.

It also disables unit tests that check for serialization error returns
in debug builds because those now abort.

FWIW this was motivated by debugging DNO-157.

Change-Id: Ia4cade82bf68d547ca914f35e1757248280014f9