[blobstore] Fix race condition between tests on teardown
To access the "fast, fifo-based block I/O" protocol,
filesystems must register a FIFO with the underlying block layer.
If the filesystem process exits (on unmount) the block device
proceeds to teardown this transaction-managing state, eventually
allowing a new process to register a FIFO.

This process is somewhat racy; if a "new" process tries registering a
FIFO with the underlying block device before it has removed the "old"
filesystem connection, then the block device may reject the "new"
process. This case is considered an "unclean shutdown".

To avoid this case, the block server offers a "clean shutdown"
alternative, encapsulated in "ioctl_block_fifo_close". During
both MinFS's and Blobstore's destructor, this "clean shutdown" is
executed, guaranteeing that the next request for a fifo (to
the underlying block server) will succeed.

This patch ensures that Blobstore actually calls its destructor
before exiting when unmounted.

Change-Id: I46a13292fa4b4b8a94d20d872da5969302d371d8