[fidl] Tweak the type of a cast in bounds checking
When zx_handle_t is changed from int32_t to uint32_t, this fails to
compile. Instead, just name the zx_handle_t type, rather than the
directly sized type, to be agnostic as to which it currently is.

This also means changing the logic to claim a handle in an encoded
message.

Change-Id: I909a0777b061f0f05accc0f18577d041df0264f5