[fvm] Provide target-side functionality for streaming sparse FVM to disk
- If a "--fvm" file is provided to the bootserver, it will be
transferred to the target device and paved onto disk. Similarly, if an
"--efi" provided to the bootserver, it will be paved onto a GPT
partition (separately from the first partition in the GPT, to avoid
dual-booting issues).
- Netsvc has been augmented so the tftp protocol can interpret
FVM and EFI files separately from netboot files, and appropriately
stream them to the FVM installer as they arrive.
- "install-disk-image" is a new application, capable of reading from
stdin (which is supplied from netsvc as a pipe) which parses the
images, and writes them to disk.

ZX-1194 #done

Change-Id: Idd6b05d7b4ff8cff81635747cc90e724f2e342d1