[sdk] Add aarch64 kernel, bootdata, and client library to SDK
This adds the qemu-aarch64 kernel and bootdata along with aarch64
client libraries to the SDK so it can be used for developing programs
that target this architecture. Chromium will use this for porting
things to aarch64 targets. This shuffles the layout of the kernel and
bootdata around, from being in:

kernel/zircon.bin
bootdata.bin

to being in

target/{x86_64,aarch64}/zircon.bin
target/{x86_64,aarch64}/bootdata.bin

This also puts the kernel debug info in
sysroot/{x86_64,aarch64}-fuchsia/debug-info to be more consistent with
other libraries.

Change-Id: Ic5a2932e2b8f671367f8168fdb4274b167291113