[getting_source] Instruct folks to download script from googlesource.com, not github
We shouldn't be telling people to download an execute a script from
github that is a mirror of the repository we actually maintain, we
should tell them to download the script from our servers (where they
will actually download everything else from). The only practical
difference between googlesource hosting and github is base64 encoding
over the wire, which is easy enough to undo on the client side.

Change-Id: Icecad428219794a9b5ffbc667e0f9228315910ff