[netstack] Put netconfig.h in public header location
This puts the netconfig.h header which defines some broadly(ish) used
IOCTLS in the garnet public header location so it can be used by other
layers without violating quite as many layering rules. This leaves
behind a forwarding header at the previous location as the Dart VM is
including this directly by path. This also changes from the
non-standard #pragma once to an old-skool include guard as some SDK
consumers are unable to process the pragma.

Change-Id: Idc0258cef992e5bb99820842b2f7328a89f548a8