Add an alias for gen.py's --modules flag.
Not long ago, we were in a situation where this was a valid command:
    packages/gn/gen.py --modules topaz/packages/modules
Even worse, the JSON in topaz/packages/modules had a field named "packages".

I'm proposing that we use the --packages flag since it points to stuff in */packages directories. Modules are something different on our system.
I'm aware of the fact that these packages are really collections of packages. I would be happy to rename these collections to something else: bundles, giftboxes, cucumbers. Anything that's not too overloaded, really.

Change-Id: Ifd4b755aa0e6c0ce43d9efa5d8bffffb339596c4