[devcoord][i915] introduce devmgr.require-system and fallback drivers
This is another (tiny) step toward driver binding priorities.

Now drivers with version strings starting with '*' are considered
"fallback drivers" which should only be loaded if we're running
a standalone zircon build rather than a full fuchsia build.

The option devmgr.require-system can be used to indicate that
the environment is a full fuchsia build.

Also:
- fixed driver enumeration to not complain about directories
- marked i915 driver as a fallback driver
- migrated i915 driver to using dprintf()

Change-Id: I1dc2b8f7f634c644a5d499caeec0416f3b05406f