Add simple routines for connecting to environment services
Many single-tenant programs need to connect to some environment
services and perform some operations on them. Currently, many of
these spin up a message loop and an app::ApplicationContext to do
this, which is quite excessive for connecting to environment services
which boils down to sending a message to a channel maintained in a
static in fdio. This adds a set of static routines for connecting in
code that is independent of a message loop and can be called from any
thread.

The reason app::ApplicationContext cares about a message loop at all
is because it hooks up an implementation to the outgoing services
interface. Code that wants to serve up outgoing services should
continue to use this, although it could be refactored more cleanly.

Change-Id: I1d9304dedbedada08c9f07c6e4cf5f742a3e59f9