Separate the update and install steps
This change separates the update and install steps into two different
commands and introduces the concepts of a lockfile (replacing the
original packagelist file). The update command generates lockfile while
the install command consumes the lockfile and produces the sysroot.
This allows running the update step separately from the install step;
typically the update step would be run by a developer while the install
step would be run by a bot to produce and upload the sysroot.

Change-Id: I79e3afb7f8cc02f749a92880decb315f542737d3