[audio] Change capturer binding behavior.
When creating an audio capturer in the media server, bind it to the
last plugged-in input instead of to the first input in the enumeration
list.

This is a temporary workaround so that folks playing with
speech-to-text can plug in and use a USB microphone if they want to.
Moving forward, input-to-capturer routing behavior should be covered
by TBD audio policy APIs.

Change-Id: I9a45f5c95b98ad49169b3f3f5058f7d1410c4fbb