[rust]ports] Update zx_port_wait and zx_port_queue to be more typesafe
The definitions of these calls used to take void* and const void* as
part of the migration from ports v1 to v2. This ambiguity is no longer
needed, and the calls now refer to zx_port_packet_t* instead. This
makes the rust wrappers match.

This also updates a few markdown links to syscall documentation.

Change-Id: Ib47b5e50b658e2b6fcb6d23a5376771b692b91b8