[bootserver] Fix string-processing error
Previously, bootserver claimed to transfer:
"bootdata.bin", and "zircon.binin".

"zircon.binin" isn't a file that was transferred;
instead, because of using memcpy to transfer "name"
into "path", the null terminator was excluded.

This patch uses "strncpy", along with an explicit
null character insertion (in the event that strlen(name)
is equal to PATH_MAX).

This results in bootserver claiming to transfer
"bootdata.bin" and "zircon.bin" correctly.

Change-Id: If2e4abf9d4a947572f203378327d4bd0347e421d