[ledger] Fix bug uploading the same commit multiple times.
Before uploading commits, we would test whether all objects have been
uploaded. When MarkPieceSynced operation in storage was synchronous
that condition was true only for the last of the objects being uploaded.
However, with storage operation potentially being asynchronous, this is
no longer the case and we might try to upload the same commit more than
once.

This CL fixes this bug by introducing a counter on pending objects
being handled.

Change-Id: I7c126fe3c636eb79db72ef1541cf108f8bbe2878