[manifest] Add jiri manifest
This CL moves the Zircon manifest from //manifest/zircon to //zircon/manifest.
This change will make it easier for the upper layers of the stack to import
Zircon at a fixed revision in the past because the upper layers will also see
the contents of the Zircon manifest from that point in the past as well.

Change-Id: I89480d6b4cb8b5dfb03f85f49b957820a43c6aa7