media: MTWN-39 fix AudioInput shutdown sequencing
AudioInput uses a worker thread and a related state variable. The
state is kStarted when the worker is running, kStopping to tell the
worker to terminate. The worker sets the state to kStopped when it
terminates.

We were getting faults when the worker detected an error and shut
down on its own. This was because the worker set the state to
kStopped before the main thread tried to shut it down, and the
main thread assumed it didn't need to join the worker thread
before destroying the std::thread. This CL removes that assumption
and joins the thread as long as its joinable. The kStopping state
has been removed, and the worker thread no longer changes the
state.

Change-Id: I6f6b7d8496e8065c6d84e809a79f1e66889a0808