[scenic] Get serious about display metrics.
Document precise definitions for units, metrics, and calibration.

Created a DisplayMetrics class to represent metrics in scenic.
Created a DisplayModel class to encapsulate metrics derivation.
Moved the display identification code into a separate function.

This patch lays the groundwork for a more rigorous treatment of
display metrics.  We will need to follow it up with tweaks to the
API, examples, and documentation which reflect the new definitions
and expose the new capabilities.

Some of the definitions, such as the usage model, are placeholders
pending further elaboration.  We might even rename the grid unit
to something else.

Bug: MZ-16
Change-Id: I34865e2f4b559a018394d32f1c872eaddbb71a22