media: MTWN-42 unbind immediately on RCHECK
RCHECK is a macro used by FactoryServiceBase 'products' to shut down
the product's primary fidl channel when a fatal error occurs. RCHECK
defers the call to UnbindAndReleaseFromOwner, because RCHECK is often
used in product constructors, and UnbindAndReleaseFromOwner cannot be
called from a product constructor (because it calls shared_from_this
and because it removes the product from the products_ list, to which
the product hasn't yet been added). The 'unbind' part of
UnbindAndReleaseFromOwner needs to happen immediately, however, to
prevent pending fidl method calls from occurring. This CL makes that
change.

Change-Id: I89b0992cd606fe889f0e2354f33f4f7079aa37bd