[manifest] Remove githook attr from manifest
As we no longer have manifest project for ziron only checkout so
removing githook attr. Right now we don't need it as the hook it
contains is already downloaded by jiri separately. But if we decide to
add more hooks we will need to create some other project or figure out
some other way to add hooks to projects.

Change-Id: Ief91f081823ca253fcdfc9a3b225d668278e945e