[manifest] Move benchmark and gflags from Zircon's manifest to Garnet's
The libraries third_party/benchmark and third_party/gflags are
dependencies of zircon_benchmarks, which has been moved to the Garnet
repo.  The dependencies should therefore be declared in Garnet's
manifest file rather than in Zircon's manifest file.

The Zircon repo itself does not depend on either of these two
libraries, because they use the C++ standard library (which is against
Zircon's policy for C++ code).

Change-Id: I4e5332c91eb440c9827f1c5b247735362a6da95f