[kernel][lib][console] Simplify cmd data structure
The only commands are the static ones, so console_register_commands
to dynamically add blocks was really not needed.  Then cmd_block is
not needed, and there is just the single linear list of cmd structs
directly in the "commands" section.

This means that section can be read-only.  Having a writable
"orphan" section is actually broken with the LLD link of the kernel
and we hadn't noticed, but this works around that problem.  (It's
not really an LLD bug, because the way we are using a dummy section
to define __data_end in the linker script is not well-defined and
it's mostly luck that the BFD linker places the dummy after all the
orphans rather than before them as LLD does.)

Change-Id: I8d219e8261bc4b6d9cf8fca3821b70d914962579